refactor, fix cyclic import
